Fix warnings and build errors in combination with inet 3.6.6 and omnet 5.5.1

+++ src/openflow/openflow/controller/OF_Controller.cc
@@ -113,7 +113,7 @@ void OF_Controller::handleMessage(cMessage *msg){
         }
 
         calcAvgQueueSize(msgList.size());
-        emit(queueSize,msgList.size());
+        emit(queueSize,static_cast<unsigned long>(msgList.size()));
     }
 }
